+++ multicam/core/services/__init__.py
@@ -4,6 +4,7 @@ from .alignment import AlignmentService, FrozenFrameInfo
 from .camera_orientation import CameraOrientationStore
 from .focus import FocusSample, FocusService
 from .dsp import DspService, DspStatus
+from .mtf import MtfFrozenFrame, MtfService
 
 __all__ = [
     "AlignmentService",
@@ -15,4 +16,6 @@ __all__ = [
     "LiveViewService",
     "DspService",
     "DspStatus",
+    "MtfFrozenFrame",
+    "MtfService",
 ]